Test for threads moved from test_window to test_thread.
