+++ fury/tests/test_thread.py
@@ -1,0 +1,55 @@
+
+import time
+from threading import Thread
+from fury.utils import rotate, update_actor, vertices_from_actor
+import numpy as np
+import numpy.testing as npt
+import pytest
+from fury import actor, window
+
+
+def test_multithreading():
+    xyz = 10 * (np.random.rand(100, 3)-0.5)
+    colors = np.random.rand(100, 4)
+    radii = np.random.rand(100) + 0.5
+
+    scene = window.Scene()
+    sphere_actor = actor.sphere(centers=xyz,
+                                colors=colors,
+                                radii=radii)
+    scene.add(sphere_actor)
+
+    # Preparing the show manager as usual
+    showm = window.ShowManager(scene,
+                            size=(900, 768), reset_camera=False,
+                            order_transparent=True)
+
+    showm.initialize()
+
+    vsa = vertices_from_actor(sphere_actor)
+
+    def callback1():
+        for i in range(100):
+
+            if(showm.lock_current()):
+                # scene.azimuth(0.01 * i)
+                rotate(sphere_actor, rotation=(0.01 * i, 1, 0, 0))
+                vsa[:] = 1.01 * vsa[:]
+                update_actor(sphere_actor)
+                print(i)
+                showm.release_current()
+                time.sleep(0.01)
+            else:
+                break
+
+        if not showm.is_done():
+            arr = window.snapshot(scene,render_window=showm.window,fname="test.png")
+            showm.exit()
+            npt.assert_equal(np.sum(arr) > 1, True)
+
+
+    thread_a = Thread(target=callback1)
+    thread_a.start()
+
+    showm.start(multithreaded=True)
+    thread_a.join()
